requirements.txt
opencv-python>=4.5.0
mediapipe>=0.10.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
pyttsx3>=2.90
pyspellchecker>=0.7.0
